+++ src/Audit.Mvc/AuditHelper.Core.cs
@@ -1,6 +1,7 @@
 ﻿#if NETCOREAPP3_0 || NETSTANDARD2_1 || NETSTANDARD2_0 || NETSTANDARD1_6 || NET451
 using Audit.Core;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.Extensions.Primitives;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -37,6 +38,20 @@ namespace Audit.Mvc
                 }
             }
             return dict.Count > 0 ? dict : null;
+        }
+
+        public static IDictionary<string, string> ToDictionary(IEnumerable<KeyValuePair<string, StringValues>> col)
+        {
+            if (col == null)
+            {
+                return null;
+            }
+            IDictionary<string, string> dict = new Dictionary<string, string>();
+            foreach (var k in col)
+            {
+                dict.Add(k.Key, string.Join(", ", k.Value));
+            }
+            return dict;
         }
     }
 }