+++ src/Audit.Mvc/ControllerExtensions.Core.cs
@@ -10,10 +10,22 @@ namespace Audit.Mvc
         /// </summary>
         /// <param name="controller">The MVC controller.</param>
         /// <returns>The current Audit Scope or NULL.</returns>
-        public static AuditScope GetCurrentAuditScope(this Microsoft.AspNetCore.Mvc.Controller controller)
+        public static AuditScope GetCurrentAuditScope(this Microsoft.AspNetCore.Mvc.ControllerBase controller)
         {
             return AuditAttribute.GetCurrentScope(controller.HttpContext);
         }
+
+#if NETCOREAPP3_0 || NETSTANDARD2_1 || NETSTANDARD2_0
+        /// <summary>
+        /// Gets the current Audit Scope.
+        /// </summary>
+        /// <param name="page">The razor page.</param>
+        /// <returns>The current Audit Scope or NULL.</returns>
+        public static AuditScope GetCurrentAuditScope(this Microsoft.AspNetCore.Mvc.RazorPages.PageModel page)
+        {
+            return AuditAttribute.GetCurrentScope(page.HttpContext);
+        }
+#endif
     }
 }
 #endif